Refactor steps into completely different algorithm in a new class
